fix narrow conversion issue

+++ cpp_beginer_to_advanced/237_custom_type_conversions/number.h
@@ -30,7 +30,9 @@ public:
 
     operator Point() const {
         std::cout << "Using the member Point() to convert\n";
-        return Point{m_wrappered_int, m_wrappered_int, m_wrappered_int};
+        return Point{static_cast<double>(m_wrappered_int),
+             static_cast<double>(m_wrappered_int),
+             m_wrappered_int};
     }
 private:
     int m_wrappered_int{};
